Added statefulness to left and right clicking
Player can left click which by default allows for unit selection, but upon changing states, will also change UI and the functionality of left click for moving a unit, aiming an attack, etc.

Right click uses a stack which just pops off the state we're at to return to the previous state the player was in. This is restricted so the player cannot pop everything off the state and return to the state prior to combat.