Added main attack actions for game loop
When choosing to do an attack, the UI is updated to show differing UI that would show the exchange of the two units.

Attacks always end up at resolve damage function inside of our unit actors, which now destroy themselves and clear references to that particular object when their health drops below a particular threshold.