Added unit deployment functionality, cleaned up a variety of things
Also replaced various SetState calls with a playercontroller SetCurrentState. This was because there were multiple instances were a whole new pointer had to be declared locally to access the state. This way, instead of having to cast to playercontroller then cast to playerstate, just cast to playercontroller. quick, easy, and simple.

Had to move and adjust a few systems revolving around spawning and placing units to accommodate for extra units.

Moved NextPlayerTurn out of pawns beginplay and into StartFight in PreBattleOptions